Nuker fixed
I may have fixed nuker for the time being.  It will notify the user that they broke a block too fast but if think about it the code that is written only allows for one block to be broken every tick.  Any more and it will send a message.  While digging by hand may work for this you will need to factor in tools and other variables.  I know what I'll be doing this weekend